pyproject.toml
[project]
name = "cs28-api"
version = "0.1.0"
description = "CS28 FastAPI backend (MySQL 5.7 + Redis)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.1",
  "pydantic==2.8.2",
  "SQLAlchemy==2.0.31",
  "aiomysql==0.2.0",
  "alembic==1.13.2",
  "python-dotenv==1.0.1",
  "PyJWT==2.8.0",
  "passlib[bcrypt]==1.7.4",
  "redis==5.0.7",
  "APScheduler==3.10.4",
  "httpx==0.27.0",
  "pytz==2024.1",
]

[tool.uvicorn]
factory = false
